+++ dst/app/routes/organizations.js
@@ -1,9 +1,4 @@
 "use strict";
-/**
- * 組織ルーター
- *
- * @ignore
- */
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
@@ -13,19 +8,37 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
     });
 };
 Object.defineProperty(exports, "__esModule", { value: true });
+/**
+ * 組織ルーター
+ */
 const express_1 = require("express");
 const organizationsRouter = express_1.Router();
 const sskts = require("@motionpicture/sskts-domain");
+const mongoose = require("mongoose");
 const authentication_1 = require("../middlewares/authentication");
 const permitScopes_1 = require("../middlewares/permitScopes");
 const validator_1 = require("../middlewares/validator");
 organizationsRouter.use(authentication_1.default);
 organizationsRouter.get('/movieTheater/:branchCode', permitScopes_1.default(['aws.cognito.signin.user.admin', 'organizations', 'organizations.read-only']), validator_1.default, (req, res, next) => __awaiter(this, void 0, void 0, function* () {
     try {
-        const repository = new sskts.repository.Organization(sskts.mongoose.connection);
-        yield repository.findMovieTheaterByBranchCode(req.params.branchCode).then((movieTheater) => {
-            res.json(movieTheater);
+        const sellerRepo = new sskts.repository.Seller(mongoose.connection);
+        const movieTheaters = yield sellerRepo.search({
+            location: { branchCodes: [req.params.branchCode] }
         });
+        const movieTheater = movieTheaters.shift();
+        if (movieTheater === undefined) {
+            throw new sskts.factory.errors.NotFound('Organization');
+        }
+        // 互換性維持のためgmoInfoをpaymentAcceptedから情報追加
+        if (Array.isArray(movieTheater.paymentAccepted)) {
+            const creditCardPaymentAccepted = movieTheater.paymentAccepted.find((p) => {
+                return p.paymentMethodType === sskts.factory.paymentMethodType.CreditCard;
+            });
+            if (creditCardPaymentAccepted !== undefined) {
+                movieTheater.gmoInfo = creditCardPaymentAccepted.gmoInfo;
+            }
+        }
+        res.json(movieTheater);
     }
     catch (error) {
         next(error);
@@ -33,10 +46,20 @@ organizationsRouter.get('/movieTheater/:branchCode', permitScopes_1.default(['aw
 }));
 organizationsRouter.get('/movieTheater', permitScopes_1.default(['aws.cognito.signin.user.admin', 'organizations', 'organizations.read-only']), validator_1.default, (__, res, next) => __awaiter(this, void 0, void 0, function* () {
     try {
-        const repository = new sskts.repository.Organization(sskts.mongoose.connection);
-        yield repository.searchMovieTheaters({}).then((movieTheaters) => {
-            res.json(movieTheaters);
+        const repository = new sskts.repository.Seller(mongoose.connection);
+        const movieTheaters = yield repository.search({});
+        movieTheaters.forEach((movieTheater) => {
+            // 互換性維持のためgmoInfoをpaymentAcceptedから情報追加
+            if (Array.isArray(movieTheater.paymentAccepted)) {
+                const creditCardPaymentAccepted = movieTheater.paymentAccepted.find((p) => {
+                    return p.paymentMethodType === sskts.factory.paymentMethodType.CreditCard;
+                });
+                if (creditCardPaymentAccepted !== undefined) {
+                    movieTheater.gmoInfo = creditCardPaymentAccepted.gmoInfo;
+                }
+            }
         });
+        res.json(movieTheaters);
     }
     catch (error) {
         next(error);