+++ dst/app/routes/people/me.js
@@ -14,8 +14,10 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const sskts = require("@motionpicture/sskts-domain");
 const createDebug = require("debug");
 const express_1 = require("express");
+const google_libphonenumber_1 = require("google-libphonenumber");
 const http_status_1 = require("http-status");
 const moment = require("moment");
+const mongoose = require("mongoose");
 const orders_1 = require("./me/orders");
 const profile_1 = require("./me/profile");
 const authentication_1 = require("../../middlewares/authentication");
@@ -52,6 +54,10 @@ meRouter.get('/contacts', permitScopes_1.default(['aws.cognito.signin.user.admin
     try {
         const personRepo = new sskts.repository.Person(cognitoIdentityServiceProvider);
         const contact = yield personRepo.getUserAttributesByAccessToken(req.accessToken);
+        // format a phone number to a Japanese style
+        const phoneUtil = google_libphonenumber_1.PhoneNumberUtil.getInstance();
+        const phoneNumber = phoneUtil.parse(contact.telephone, 'JP');
+        contact.telephone = phoneUtil.format(phoneNumber, google_libphonenumber_1.PhoneNumberFormat.NATIONAL);
         res.json(contact);
     }
     catch (error) {
@@ -62,21 +68,29 @@ meRouter.get('/contacts', permitScopes_1.default(['aws.cognito.signin.user.admin
  * 会員プロフィール更新
  * @deprecated
  */
-meRouter.put('/contacts', permitScopes_1.default(['aws.cognito.signin.user.admin', 'people.contacts']), (__1, __2, next) => {
-    next();
-}, validator_1.default, (req, res, next) => __awaiter(this, void 0, void 0, function* () {
+meRouter.put('/contacts', permitScopes_1.default(['aws.cognito.signin.user.admin', 'people.contacts']), validator_1.default, (req, res, next) => __awaiter(this, void 0, void 0, function* () {
     try {
-        const personRepo = new sskts.repository.Person(cognitoIdentityServiceProvider);
-        yield personRepo.updateContactByAccessToken({
-            accessToken: req.accessToken,
-            contact: {
-                givenName: req.body.givenName,
-                familyName: req.body.familyName,
-                email: req.body.email,
-                telephone: req.body.telephone
+        // 日本語フォーマットで電話番号が渡される想定なので変換
+        let formatedPhoneNumber;
+        try {
+            const phoneUtil = google_libphonenumber_1.PhoneNumberUtil.getInstance();
+            const phoneNumber = phoneUtil.parse(req.body.telephone, 'JP');
+            if (!phoneUtil.isValidNumber(phoneNumber)) {
+                throw new Error('Invalid phone number format.');
             }
+            formatedPhoneNumber = phoneUtil.format(phoneNumber, google_libphonenumber_1.PhoneNumberFormat.E164);
+        }
+        catch (error) {
+            next(new sskts.factory.errors.Argument('telephone', 'invalid phone number format'));
+            return;
+        }
+        const personRepo = new sskts.repository.Person(cognitoIdentityServiceProvider);
+        yield personRepo.updateProfileByAccessToken({
+            accessToken: req.accessToken,
+            profile: Object.assign({}, req.body, { telephone: formatedPhoneNumber })
         });
-        res.status(http_status_1.NO_CONTENT).end();
+        res.status(http_status_1.NO_CONTENT)
+            .end();
     }
     catch (error) {
         next(error);
@@ -141,7 +155,7 @@ meRouter.post('/accounts', permitScopes_1.default(['aws.cognito.signin.user.admi
             accountNumber: accountNumberRepo,
             accountService: accountService
         });
-        const ownershipInfoRepo = new sskts.repository.OwnershipInfo(sskts.mongoose.connection);
+        const ownershipInfoRepo = new sskts.repository.OwnershipInfo(mongoose.connection);
         const ownershipInfo = {
             typeOf: 'OwnershipInfo',
             // 十分にユニーク
@@ -156,7 +170,7 @@ meRouter.post('/accounts', permitScopes_1.default(['aws.cognito.signin.user.admi
             // tslint:disable-next-line:no-magic-numbers
             ownedThrough: moment(now).add(100, 'years').toDate() // 十分に無期限
         };
-        yield ownershipInfoRepo.save(ownershipInfo);
+        yield ownershipInfoRepo.saveByIdentifier(ownershipInfo);
         res.status(http_status_1.CREATED).json(account);
     }
     catch (error) {
@@ -170,8 +184,8 @@ meRouter.post('/accounts', permitScopes_1.default(['aws.cognito.signin.user.admi
 meRouter.put('/accounts/:accountNumber/close', permitScopes_1.default(['aws.cognito.signin.user.admin', 'people.accounts']), validator_1.default, (req, res, next) => __awaiter(this, void 0, void 0, function* () {
     try {
         // 口座所有権を検索
-        const ownershipInfoRepo = new sskts.repository.OwnershipInfo(sskts.mongoose.connection);
-        const accountOwnershipInfos = yield ownershipInfoRepo.search({
+        const ownershipInfoRepo = new sskts.repository.OwnershipInfo(mongoose.connection);
+        const accountOwnershipInfos = yield ownershipInfoRepo.search4cinemasunshine({
             goodType: sskts.factory.pecorino.account.TypeOf.Account,
             ownedBy: req.user.username
         });
@@ -224,8 +238,8 @@ meRouter.delete('/accounts/:accountNumber', permitScopes_1.default(['aws.cognito
     try {
         const now = new Date();
         // 口座所有権を検索
-        const ownershipInfoRepo = new sskts.repository.OwnershipInfo(sskts.mongoose.connection);
-        const accountOwnershipInfos = yield ownershipInfoRepo.search({
+        const ownershipInfoRepo = new sskts.repository.OwnershipInfo(mongoose.connection);
+        const accountOwnershipInfos = yield ownershipInfoRepo.search4cinemasunshine({
             goodType: sskts.factory.pecorino.account.TypeOf.Account,
             ownedBy: req.user.username,
             ownedAt: now
@@ -252,8 +266,8 @@ meRouter.get('/accounts', permitScopes_1.default(['aws.cognito.signin.user.admin
             throw new sskts.factory.errors.Forbidden('Login required');
         }
         // 口座所有権を検索
-        const ownershipInfoRepo = new sskts.repository.OwnershipInfo(sskts.mongoose.connection);
-        const accountOwnershipInfos = yield ownershipInfoRepo.search({
+        const ownershipInfoRepo = new sskts.repository.OwnershipInfo(mongoose.connection);
+        const accountOwnershipInfos = yield ownershipInfoRepo.search4cinemasunshine({
             goodType: sskts.factory.pecorino.account.TypeOf.Account,
             ownedBy: req.user.username,
             ownedAt: now
@@ -284,8 +298,8 @@ meRouter.get('/accounts/:accountNumber/actions/moneyTransfer', permitScopes_1.de
     try {
         const now = new Date();
         // 口座所有権を検索
-        const ownershipInfoRepo = new sskts.repository.OwnershipInfo(sskts.mongoose.connection);
-        const accountOwnershipInfos = yield ownershipInfoRepo.search({
+        const ownershipInfoRepo = new sskts.repository.OwnershipInfo(mongoose.connection);
+        const accountOwnershipInfos = yield ownershipInfoRepo.search4cinemasunshine({
             goodType: sskts.factory.pecorino.account.TypeOf.Account,
             ownedBy: req.user.username,
             ownedAt: now
@@ -315,8 +329,8 @@ meRouter.get('/ownershipInfos/:goodType', permitScopes_1.default(['aws.cognito.s
     next();
 }, validator_1.default, (req, res, next) => __awaiter(this, void 0, void 0, function* () {
     try {
-        const repository = new sskts.repository.OwnershipInfo(sskts.mongoose.connection);
-        const ownershipInfos = yield repository.search({
+        const repository = new sskts.repository.OwnershipInfo(mongoose.connection);
+        const ownershipInfos = yield repository.search4cinemasunshine({
             goodType: req.params.goodType,
             ownedBy: req.user.username,
             ownedAt: new Date()
@@ -343,9 +357,9 @@ meRouter.put('/ownershipInfos/programMembership/register', permitScopes_1.defaul
             programMembershipId: req.body.programMembershipId,
             offerIdentifier: req.body.offerIdentifier
         })({
-            organization: new sskts.repository.Organization(sskts.mongoose.connection),
-            programMembership: new sskts.repository.ProgramMembership(sskts.mongoose.connection),
-            task: new sskts.repository.Task(sskts.mongoose.connection)
+            seller: new sskts.repository.Seller(mongoose.connection),
+            programMembership: new sskts.repository.ProgramMembership(mongoose.connection),
+            task: new sskts.repository.Task(mongoose.connection)
         });
         // 会員登録タスクとして受け入れられたのでACCEPTED
         res.status(http_status_1.ACCEPTED).json(task);
@@ -366,8 +380,8 @@ meRouter.put('/ownershipInfos/programMembership/:identifier/unRegister', permitS
             agent: req.agent,
             ownershipInfoIdentifier: req.params.identifier
         })({
-            ownershipInfo: new sskts.repository.OwnershipInfo(sskts.mongoose.connection),
-            task: new sskts.repository.Task(sskts.mongoose.connection)
+            ownershipInfo: new sskts.repository.OwnershipInfo(mongoose.connection),
+            task: new sskts.repository.Task(mongoose.connection)
         });
         // 会員登録解除タスクとして受け入れられたのでACCEPTED
         res.status(http_status_1.ACCEPTED).json(task);