+++ dst/mongooseConnectionOptions.js
@@ -3,14 +3,14 @@ Object.defineProperty(exports, "__esModule", { value: true });
 /**
  * mongoose接続オプション
  * @see http://mongoosejs.com/docs/api.html#index_Mongoose-connect
- * @ignore
  */
 const mongooseConnectionOptions = {
     autoReconnect: true,
-    keepAlive: 120000,
+    keepAlive: true,
     connectTimeoutMS: 30000,
     socketTimeoutMS: 0,
     reconnectTries: 30,
-    reconnectInterval: 1000
+    reconnectInterval: 1000,
+    useNewUrlParser: true
 };
 exports.default = mongooseConnectionOptions;